feat: implement TCS ledger persistence and tax splitting logic in checkout service
- Added `persistTcsCollectionLedger` function to handle TCS ledger entries with proper tax splitting for intra-state and inter-state transactions.
- Refactored existing TCS handling in the checkout process to utilize the new persistence function.
- Introduced unit tests for TCS ledger creation and tax splitting to ensure accuracy and regression protection.
- Updated related services and tests to accommodate the new tax calculation methods and ensure consistent behavior across the application.

+++ src/modules/reports/definitions/adminFinance.ts
@@ -435,14 +435,14 @@ async function reconciliation(filters: ReportFilters) {
     `
     SELECT
       COALESCE((
-        SELECT SUM(wro."amountInr")::float
+        SELECT SUM(wro."amountInr")::numeric
         FROM wallet_recharge_orders wro
         WHERE wro.status = 'PAID'
           AND wro."paidAt" BETWEEN :from AND :to
           AND wro."deletedAt" IS NULL
       ), 0) AS "walletRechargeInflow",
       COALESCE((
-        SELECT SUM(o."walletAmountUsed")::float
+        SELECT SUM(o."walletAmountUsed")::numeric
         FROM orders o
         WHERE o."createdAt" BETWEEN :from AND :to
           AND o."deletedAt" IS NULL
